+++ mart/attack/__init__.py
@@ -7,4 +7,5 @@ from .gain import *
 from .gradient_modifier import *
 from .initializer import *
 from .objective import *
+from .perturber import *
 from .projector import *
