+++ mart/configs/attack/adversary.yaml
@@ -1,10 +1,8 @@
 _target_: mart.attack.Adversary
-initializer: ???
+perturber: ???
 optimizer: ???
-composer: ???
 gain: ???
 gradient_modifier: null
-projector: null
 objective: null
 enforcer: ???
 attacker: null
