+++ mart/configs/attack/object_detection_mask_adversary_missed.yaml
@@ -5,8 +5,9 @@ defaults:
 
 max_iters: 100
 
-initializer:
-  constant: 127
-
 optimizer:
   lr: 3
+
+perturber:
+  initializer:
+    constant: 127
